Initial contribution.
Provides a first functional version of the SARL Agent-Oriented
Programming Language.

SARL is a general-purpose agent-oriented language.

SARL aims at providing the fundamental abstractions for dealing with
concurrency, distribution, interaction, decentralization, reactivity,
autonomy and dynamic reconfiguration.
These high-level features are now considered as the major requirements
for an easy and practical implementation of modern complex software
applications.
We are convinced that the agent-oriented paradigm holds the keys to
effectively meet this challenge.

Considering the variety of existing approaches and meta-models in the
field of agent-oriented engineering and more generally multi-agent
systems, our approach remains as generic as possible and highly
extensible to easily integrate new concepts and features. The language
is platform- and architecture-independent.

However, we provide a first set of tools to support its execution based
on the new version 2.0 of the Janus (see http://www.janusproject.io)
platform, but it can be linked with other existing agent platforms and
frameworks.