Cleaning the root pom file.
Generate the Javadoc is now generated in the "release" profile. This
avoid to take time and memory during snapshot compilation. Indeed the
javadoc are not really necessary since the source code is packaged. Ask
tthe clean maven plugin to remove the folder "workspace", which is the
result of the execution of the Eclipse for the tests.