[maven] Add the Maven plugin for docs.
Jnario generates the HTML structure according to the types of the
elements in the specification (fact, description, etc.) This causes an
invalid rendering of the documentation pages. Indeed, they are written
by assuming that the depth of the elements in the specification is
relevant to the structure of the documentation. The Maven plugin is
generated the HTML tags according to the depth of the elements.

Additionally, it also does the following replacements:

* Replace "<!-- OUTPUT OUTLINE -->" and "@outline" by the outline of the
specification.
* Replace "<note/>", "<importantnote/>", "<cautionote/>", and
"<veryimportantnote/>" by the corresponding HTML code.

Signed-off-by: Stéphane Galland <galland@arakhne.org>