[eclipse] Make the SARL libraries as provided libraries
Create a library container for compiling SARL code.
This container is a "K_DEFAULT_SYSTEM" library: it is used during the
compilation, but not added in the runtime class path. It is assumed that
the system is providing these libraries. This is the case since Janus is
providing them in its "full dependency" jar file.

see #205

Signed-off-by: Stéphane Galland <galland@arakhne.org>