[lang] Add the @FiredEvents and @EarlyExit annotations
The @FiredEvents annotation permits to mark the Java functions with the
SARL events that are specified with the "fires" keyword.

The @EarlyExit annotation is used for marking the methods that never
return. In this case, the SARL validator output a "Unreachable code"
error for any statement following the call to a marked method.

The fired "Destroy" event (marked by @FiredEvent) is used for marking
the method with @EarlyExit.

see #202

Signed-off-by: Stéphane Galland <galland@arakhne.org>