Updating the management of libraries provided by an SRE to deal with Classpath Entries instead of library locations
The goal was to be more generic in the management of SRE classpath
extensions to deal with any type of classpath entries instead of just
library location. A library location can just be created from a jar but
sometimes especially in product debug mode we need projet-based
dependencies or other kinds of classpath extensions.