Changing all Locale.getString to use a specified class
This problem avoid a problem in find the right classloader when you're
using osgi. In fact it returned the classloader of vmutil plugin instead
of the classloader of the plugin where the specified class is
effectively loaded. By specfiying the class we avoid this problem, we
are sure to obtain the right classloader.