[all] Major project folder refactoring.
In order to have the correct trace file generated from the
"io.sarl.core" and "io.sarl.util" plugins, the sarl-maven-plugin should
be used, and not the xtext-maven-plugin.

For enabling the use of the sarl-maven-plugin, the folder structure of
the SARL project should be refactoring for compiling the sarl-maven-
plugin module first, then the SDK/API modules.

close #608

Signed-off-by: Stéphane Galland <galland@arakhne.org>