[tests] Fast failure when UTBM servers are not reachable.
Several servers of `sarl.io` are hosted by UTBM.
When UTBM is disconnected from Internet or has several network
infrastructure issues, the SARL unit tests may wait for a long time
without a network timeout.

This patch tests the reachability UTBM's servers as fast as possible and
force all the unit tests that are needed one of these servers to be
ignored.

see #649

Signed-off-by: Stéphane Galland <galland@arakhne.org>