[lang] Do not break the internal state of JVM operations when detecting purability dynamically.
Before this patch, the tools that detects the purability of a function
may break the internal state of the JvmOperation object, such that the
Java-code generator cannot generate the function calls to the broken
JvmOperation. This patch changes the way how the purability of the
JvmOperation objects is detected. The Java-code generator add the
"@Pure" annotation without changing the JvmOperation object's state.

Signed-off-by: Stéphane Galland <galland@arakhne.org>