[eclipse] Refactor the validation code of the editor preference page.
The use of the validation flag in the editor preference page did not
follow the specification of the super preference page. The code is
refactored in order to ensure that the validation flag is evaluate to
false only if a value within the preference page is invalid from its
format point-of-view.

close #837

Signed-off-by: Stéphane Galland <galland@arakhne.org>