[lang] Remove the search of Xtext library on the classpath when validating.
The search method for finding the Xtext library on the classpath is
based on Xtext utility functions that provide the version of the Xtext
into the classpath of the validator, but not the classpath of the
compiler project. It is in contradiction with the test on the SARL
version, which is done from the SARL SDK on the compiled project's
classpath. There is no programmatic method for moving the Xtext
detection from the compiler classpath to the project classpath, except
changing Xtext itself.

Signed-off-by: Stéphane Galland <galland@arakhne.org>