[sarlc] Deep re-implementation in order to fix the issues of classpath.
Sarlc tools had issues related to the management of the classpath. This
patch provides: i) a new implementation of sarlc based on the new
"io.sarl.maven.bootique" module, ii) a new classpath detector that could
be used as path of a larger application (when sarlc is embedded into
another application, e.g. sarldoc) or from the command line.

Signed-off-by: Stéphane Galland <galland@arakhne.org>