[lang] Refactor the DefaultActionPrototype type in order to limit memory leaks.
Several tests indicated that DefaultActionPrototype takes 250Mb of heap
space over the 350Mb allocated by the SARL IDE.
This issue is caused by a buffer that is not clear in an appropriate
way. A context, which contains the buffer is added in order to control
the life-time of this buffer.

Signed-off-by: Stéphane Galland <galland@arakhne.org>