[all] Do not save the "devel" buildId into files.
In order to avoid the need of commiting the "about.mappings" files each
time we are launch a Maven compilation, the buildId that is saved into
the Eclipse configuration files in static other the entire lifetime of
the development version. This static buildId will be replaced by the
real (and standard) buildId into the released version.

Signed-off-by: Stéphane Galland <galland@arakhne.org>