[SRE] Sovling the bug in PingPong Demo where an event was not correctly sent due to a problem of WeakReference
The WeakRef from ExternalEventBusAccessor to InternalEventBusSkill
became null after a certain number of agent cycles. We were UNABLE TO
EXPLAIN this problem, it si clearly multithread problem because in Debug
the problem didn't appear. We decide to remove this weakreference and
manage manually the memory free but it didn't really solved the original
problem just make Janus work properly.