[maven] Split module "io.sarl.maven.bootiqueapp".
This module is used to build command-line tools. But it included
Bootique and the SARL batch compiler. In order to create tools that are
not linked to the batch compiler, the module is split into
"io.sarl.maven.bootiqueapp" and "io.sarl.maven.bootiquebatchcompiler".
In this way, the tools may depends on a specific module depends if they
need the batch compiler or not.

see #1000

Signed-off-by: Stéphane Galland <galland@arakhne.org>