[ui] Add the definition of polymorphic methods for providing the code mining contributions.
In order to be more efficient, the different providers of code mining
contributions are not run separatly because they are iterating on the
SARL program's syntax tree all. This duplication of the iteration in the
different providers is time consuming. This commit iterator a single
time on the syntax tree and polymorphic methods are invoked on the
syntax tree elements for generating the code mining contributions.

see #1072

Signed-off-by: Stéphane Galland <galland@arakhne.org>